moved HostItem to separate file
