+++ VSRAD.Package/ProjectSystem/Profiles/TargetHostsEditor.xaml.cs
@@ -12,24 +12,6 @@ namespace VSRAD.Package.ProjectSystem.Profiles
 {
     public sealed partial class TargetHostsEditor : DialogWindow
     {
-        public sealed class HostItem : DefaultNotifyPropertyChanged
-        {
-            private string _host = "";
-            public string Host { get => _host; set => SetField(ref _host, value); }
-
-            private string _alias = "";
-            public string Alias { get => _alias; set => SetField(ref _alias, value); }
-
-            public bool UsedInActiveProfile { get; }
-
-            public HostItem(string host, bool usedInActiveProfile, string alias = "")
-            {
-                Alias = alias;
-                Host = host;
-                UsedInActiveProfile = usedInActiveProfile;
-            }
-        }
-
         public static bool TryParseHost(string input, out string formatted, out string hostname, out ushort port)
         {
             formatted = "";
